Crate ibc_types_core_channel

IBC channel-related types.

Re-exports

pub use channel::ChannelEnd;

pub use channel::Counterparty;

pub use channel::IdentifiedChannelEnd;

pub use packet::Packet;

Modules

channel

events

Types for the IBC events emitted from Tendermint Websocket by the channels module.

msgs

Message definitions for all ICS4 domain types: channel open & close handshake datagrams, as well as packets.

packet

Structs

AcknowledgementCommitment

Acknowledgement commitment to be stored

ChannelId

PacketCommitment

Packet commitment

PortId

Version

The version field for a ChannelEnd.

Enums

ChannelError

PacketError

TimeoutHeight

Indicates a consensus height on the destination chain after which the packet will no longer be processed, and will instead count as having timed-out.